Simplification des prints install.sh

+++ install.py
@@ -28,9 +28,9 @@ liste_pages_a_copier = [
 env = Environment(loader=FileSystemLoader('.'))
 
 # (Re-)Création du dossier build du site
-shutil.rmtree('build')
+if os.path.exists('build'):
+    shutil.rmtree('build')
 os.makedirs('build')
-print("(Re-)Création du dossier build terminée\n")
 
 # Copie du dossier backend dans le dossier build
 shutil.copytree('backend', './build/backend')
@@ -48,13 +48,10 @@ shutil.copytree('./frontend/scss', './build/frontend/scss')
 
 os.chdir('./frontend')
 
-
-print("> Rendu des pages en cours...\n")
-
 # Effectue le rendu de chaque page
 for page in liste_pages_a_rendre:
     page = page + '.html'
-    print("\tRendu de la page '" + page + "'...")
+    print("Rendu de la page '" + page + "'...")
 
     page_path = page
     new_page_path = "../build/frontend/" + page
@@ -70,24 +67,13 @@ for page in liste_pages_a_rendre:
     fichier.write(output)
     fichier.close()
 
-    print("\tRendu terminé.\n")
-
-print("Fin de rendu des pages.\n", "\n> Copie des autres pages...\n")
-
 # Copie des pages qui n'ont pas besoin de rendu
 for page in liste_pages_a_copier:
     page = page + '.html'
-    print("\tCopie de la page '" + page + "'...")
+    print("Copie de la page '" + page + "'...")
 
     page_path = page
     new_page_path = "../build/frontend/" + page
 
     shutil.copy(page_path, new_page_path)
     
-    print("\tCopie terminée.\n")
-
-print("Fin de copie des pages.\n")
-
-
-# Fin du programme
-print('-'*32 + "\n\tRendu complet !" + "\n" + '-'*32)
